+++ packages/scss-utils/CHANGELOG.md
@@ -5,6 +5,16 @@
 
 ## [1.0.0-rc.1] (2019-12-13)
 
+* Improve neutral color namings and variations
+  * $p-color-theme-light-neutral-1 => $p-color-theme-light-neutral-contrast-high
+  * $p-color-theme-light-neutral-2 => removed without replacement
+  * $p-color-theme-light-neutral-3 => $p-color-theme-light-neutral-contrast-medium
+  * Add $p-color-theme-light-neutral-contrast-low
+  * $p-color-theme-dark-neutral-1 => $p-color-theme-dark-neutral-contrast-high
+  * $p-color-theme-dark-neutral-2 => removed without replacement
+  * $p-color-theme-dark-neutral-3 => $p-color-theme-dark-neutral-contrast-medium
+  * Add $p-color-theme-dark-neutral-contrast-low
+
 ### Features
 
 * Initial release